issue #205015 Telemetry Response Event is not capturing properly for Configurable Telemetry
